examples/fastapi-notes/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation-logging==0.47b0
